requirements.txt
Django>=4.2
djangorestframework>=3.14
requests>=2.31
beautifulsoup4>=4.12
